bg変更

+++ マテスナ/bullet.h
@@ -10,18 +10,18 @@ public:
 	void Init();
 	void Update(Aim& aim, Target& tag, float& gameTime, float& deltaTime);
 	void Draw();
-	void DrawBulletMark(float& mouseX, float&mouseY, int& objX, int& objY, float& exRate, bool& flag);
+	void DrawBulletMark(float& mouseX, float&mouseY, float& objX, float& objY, float& exRate, bool& flag);
 	bool GetImpactFlag() const{ return impactFlag; }
 private:
 	int Markhandle;
-	int x, y;
+	float x, y;
 	float z;
 	const float initialV = 8.68f;//èâë¨
 
 	bool aliveFlag;
 	bool fireFlag;
 	bool drawFlag;
-	int prevX, prevY;
+	float prevX, prevY;
 	bool impactFlag;
 };
